Moved periods outside quoted strings
The title of OMB Circular A-11 (the 2014 version of which can be
reviewed at
http://www.whitehouse.gov/sites/default/files/omb/assets/a11_current_yea
r/a11_2014.pdf) is an exact string that does not contain a period.
Hence, the period has been moved outside the quotes surrounding the
title, and an additional similar change has been made to maintain
stylistic coherence.